No longer required pre-declaring bitfields.
The code no longer has to pre-define bitfields it intends to
use which was a rather clunky way to do it.  Now we pass in
a string, declare / define a variable and assign it 0.  Then
handle it as usual.  We need to make it so we can support
arbitrary number of if statements.