Partly works.
It partly works.  It doesn't catch the case where (in the airport event
handler) we do:

s->furthest_flight_landing += ts;

This is bad because ts is a local variable that we load.  We need to
make sure that the intersection of all loads hits something within the
target argument.